Phase1/ML/Features-Builder

+++ db/views.sql
@@ -1,5 +1,5 @@
 -- ============================================================
--- DuctAI Copilot — Analytical Views
+-- DuctAI Copilot - Analytical Views
 -- Run against a live database: psql $DATABASE_URL -f db/views.sql
 -- Safe to re-run: uses CREATE OR REPLACE VIEW
 --
@@ -15,9 +15,9 @@
 --     8.  v_customer_retention
 --     9.  v_service_cooccurrence
 --   ML support (10-12):
---     10. v_customer_identity    — deduplicates returning customers by phone
---     11. v_customer_history     — per-person job history using canonical ID
---     12. v_city_price_index     — median job value per city for pricing model
+--     10. v_customer_identity    - deduplicates returning customers by phone
+--     11. v_customer_history     - per-person job history using canonical ID
+--     12. v_city_price_index     - median job value per city for pricing model
 -- ============================================================
 
 
@@ -188,8 +188,8 @@ ORDER BY total_revenue DESC NULLS LAST;
 -- ------------------------------------------------------------
 -- 7. Estimate pipeline
 -- Q: What is the open revenue pipeline and conversion rate?
--- Uses consolidated `value` column (won/lost/open merged in ETL).
--- Two conversion rates — see data_notes.md for explanation.
+-- Uses consolidated value column (won/lost/open merged in ETL).
+-- Two conversion rates - see data_notes.md for explanation.
 -- ------------------------------------------------------------
 CREATE OR REPLACE VIEW v_estimate_pipeline AS
 SELECT
@@ -197,12 +197,10 @@ SELECT
     COUNT(*) FILTER (WHERE outcome = 'won')                 AS won_count,
     COUNT(*) FILTER (WHERE outcome = 'lost')                AS lost_count,
     COUNT(*) FILTER (WHERE outcome = 'open')                AS open_count,
-    -- Decision conversion: won / (won + lost) — excludes still-open
     ROUND(
         COUNT(*) FILTER (WHERE outcome = 'won') * 100.0 /
         NULLIF(COUNT(*) FILTER (WHERE outcome IN ('won','lost')), 0)
     , 1)                                                    AS decision_conversion_pct,
-    -- Overall conversion: won / all estimates including open
     ROUND(
         COUNT(*) FILTER (WHERE outcome = 'won') * 100.0 /
         NULLIF(COUNT(*), 0)
@@ -217,7 +215,7 @@ FROM estimates;
 -- ------------------------------------------------------------
 -- 8. Customer retention segments
 -- Q: What share of revenue comes from repeat customers?
--- Note: uses customer_id directly — may undercount true returning
+-- Note: uses customer_id directly - may undercount true returning
 -- customers due to Jobber duplicate records. See v_customer_identity.
 -- ------------------------------------------------------------
 CREATE OR REPLACE VIEW v_customer_retention AS
@@ -277,14 +275,8 @@ ORDER BY co_count DESC;
 -- Problem: Jobber creates a new customer_id for returning
 -- customers booked via pipeline automation or Thumbtack imports.
 -- 81 phone numbers are linked to 2+ customer IDs.
--- 47 are confirmed returning customers with new records.
--- 21 are double-entry errors (created within minutes of each other).
---
 -- Solution: canonical_id = earliest ID sharing the same phone.
 -- This is the true person identifier for all ML features.
---
--- Usage: always JOIN through this view before aggregating
--- job history for ML feature building.
 -- ------------------------------------------------------------
 CREATE OR REPLACE VIEW v_customer_identity AS
 SELECT
@@ -316,14 +308,6 @@ WHERE c.name NOT ILIKE '%express job%'
 -- Aggregates all jobs per canonical person, not per record ID.
 -- Correctly counts job history for returning customers who
 -- were re-created in Jobber with a new customer ID.
---
--- ML features:
---   prior_job_count       — total completed jobs as a person
---   prior_avg_job_value   — average spend across all their jobs
---   prior_max_job_value   — highest single job (upsell ceiling)
---   is_returning_customer — has at least one prior completed job
---   last_job_date         — recency signal
---   lead_source           — fills the 97% null gap in estimates
 -- ------------------------------------------------------------
 CREATE OR REPLACE VIEW v_customer_history AS
 SELECT
@@ -348,30 +332,18 @@ GROUP BY ci.canonical_id;
 -- 12. City price index
 -- Pre-computes median and percentile job values per city.
 -- Used by the pricing model as a market price signal.
--- Cities with fewer than 3 completed jobs are excluded —
--- insufficient data for a reliable median.
--- In the ML pipeline, low-sample cities fall back to state median.
---
--- ML features:
---   median_job_value — central price for this market
---   p25_job_value    — lower bound of typical range
---   p75_job_value    — upper bound of typical range
+-- Cities with fewer than 3 completed jobs are excluded.
+-- Low-sample cities fall back to state median in ML pipeline.
 -- ------------------------------------------------------------
 CREATE OR REPLACE VIEW v_city_price_index AS
 SELECT
     city,
     state,
-    COUNT(id)                                   AS job_count,
-    ROUND(
-        PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY job_amount)
-    , 2)                                        AS median_job_value,
-    ROUND(AVG(job_amount), 2)                   AS avg_job_value,
-    ROUND(
-        PERCENTILE_CONT(0.25) WITHIN GROUP (ORDER BY job_amount)
-    , 2)                                        AS p25_job_value,
-    ROUND(
-        PERCENTILE_CONT(0.75) WITHIN GROUP (ORDER BY job_amount)
-    , 2)                                        AS p75_job_value
+    COUNT(id)                                       AS job_count,
+    ROUND(PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY job_amount)::NUMERIC, 2)  AS median_job_value,
+    ROUND(AVG(job_amount), 2)                       AS avg_job_value,
+    ROUND(PERCENTILE_CONT(0.25) WITHIN GROUP (ORDER BY job_amount)::NUMERIC, 2) AS p25_job_value,
+    ROUND(PERCENTILE_CONT(0.75) WITHIN GROUP (ORDER BY job_amount)::NUMERIC, 2) AS p75_job_value
 FROM jobs
 WHERE status = 'Completed'
   AND job_amount > 0
